Refactor to avoid use of kubectl proxy
kubectl proxy was shown to cause watch updates to get reordered.  This was
causing scheduling queries to return empty because the expected data was
missing.  E.g., unscheduled pod updates were arriving before the corresponding
replication controller update.

In the future, we should refactor the scheduler to use the upcoming Kubernetes
go-client.